#9: try-finally for workflows
Implemented a try-finally structure for the workflow. It is now possible
to embed a whole workflow into a try block. The corresponding finally
block will always be executed, even if the previous workflow of the try
block needed to be rolled back. The finally workflow also supports the
rollback mechanism in case of a failure.